+++ axum-macros/tests/from_request/pass/named_parts.rs
@@ -1,7 +1,11 @@
 use axum::{
-    extract::{rejection::TypedHeaderRejection, FromRequestParts, TypedHeader},
-    headers::{self, UserAgent},
+    extract::FromRequestParts,
     response::Response,
+};
+use axum_extra::{
+    TypedHeader,
+    typed_header::TypedHeaderRejection,
+    headers::{self, UserAgent},
 };
 
 #[derive(FromRequestParts)]