+++ axum/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+- None.
+
+# 0.6.17 (25. April, 2023)
+
 - **fixed:** Fix fallbacks causing a panic on `CONNECT` requests ([#1958])
 
 [#1958]: https://github.com/tokio-rs/axum/pull/1958